pyproject.toml
[tool.poetry]
name = "finnancia-fastapi"
version = "0.1.0"
description = ""
authors = ["Enzo Bozzani <enzobozzani@ibm.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.0"}
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
sqlmodel = "^0.0.21"
pyjwt = "^2.9.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
types-passlib = "^1.7.7.20240327"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
